requirements.txt
streamlit==1.44.0
requests==2.32.3
beautifulsoup4==4.13.3
googlesearch-python==1.3.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
google-generativeai==0.8.4
reportlab==4.3.1
python-dotenv